Security Group Hook was moved to separate file
